users/requirements.txt
Flask==2.2.2
flask-jwt-extended==4.3.1
supabase==2.0.2
python-dotenv==0.19.2
Werkzeug==2.2.2
